Struct mentat_sql::SQLQuery

pub struct SQLQuery {
    pub sql: String,
    pub args: Vec<(String, Rc<Value>)>,
}

We want to accumulate values that will later be substituted into a SQL statement execution. This struct encapsulates the generated string and the initial argument list. Additional user-supplied argument bindings, with their placeholders accumulated via push_bind_param, will be appended to this argument list.

Fields
sql: String args: Vec<(String, Rc<Value>)>

These will eventually perhaps be rusqlite ToSql instances.